JS: fix `is PrimitiveArray` and prepare to map primitive arrays to     TypedArray's (KT-15358, KT-14007, KT-16056, KT-14614 fixed)
Invariant: for compatibility reasons at any point a PrimitiveArray
may be represented by either a regular array, or a TypedArray, with
or without the $type$ property.

Move TypedArray's and is-checks for PrimitiveArray's behind -Xtypedarrays

Added a compilation key '-Xtypedarrays' which enables translation to
TypedArray's and is-checks for primitive arrays.

Added box tests for codegen/box/arrays with the key enabled.

Rewrote the primitiveArrays.kt test so that it works with the
key disabled,

Tweaked runtime, so that it works when compiled with the flag
disabled.